+++ modules/versions.lua
@@ -6,7 +6,7 @@ return {
   ["Gen151"] = "1.5.1",
   ["Gen1AutoContinue"] = "1.2.2",
   ["Gen1AutoSave"] = "1.4.0",
-  ["Gen1Follower"] = "1.2.1",
+  ["Gen1Follower"] = "1.3.1",
   ["Gen1MenuManager"] = "0.2.8",
   ["Gen1ModMenu"] = "0.9.0",
   ["Gen1Remember"] = "1.0.1",